pyproject.toml
[project]
name = "celery-prefork-safety"
version = "0.1.0"
description = "Demonstration of a fork-safety bug with ChromaDB and Celery's prefork worker pool."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery==5.5.3",
    "eventlet>=0.40.3",
    "fastapi>=0.121.2",
    "gevent>=25.9.1",
    "langchain==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-community==0.3.30",
    "langchain-core==0.3.77",
    "langchain-google-genai==2.1.12",
    "langchain-openai>=0.3.34",
    "langchain-text-splitters==0.3.11",
    "langsmith==0.4.31",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Target Python 3.13 (matches project requires-python)
target-version = "py313"

[tool.ruff.lint]
# Essential rules for Django + pytest project
select = [
  "F",      # pyflakes - undefined names, unused imports/variables
  "E",      # pycodestyle errors - syntax errors, indentation (all E errors, not just E4/E7/E9)
  "W",      # pycodestyle warnings
  "I",      # isort - import sorting
  "N",      # pep8-naming - naming conventions
  "UP",     # pyupgrade - update old Python syntax
  "B",      # flake8-bugbear - catches common bugs
  "DJ",     # flake8-django - Django-specific checks
  "PT",     # flake8-pytest-style - pytest best practices
  "D",      # pydocstyle - docstring checks (Google style)
  "C90",    # mccabe - complexity checks
  "ERA001", # Found commented-out code
  "A",      # flake8-builtins
  "C4",     # flake8-comprehensions
  "DTZ",    # flake8-datetimez
  "FA",     # flake8-future-annotations
  "INT",    #flake8-gettext
  "ICN",    # flake8-import-conventions
  "G",      # flake8-logging-format
  "PIE",    # flake8-pie
  "RET",    # flake8-return
  "ARG",    # flake8-unused-arguments
  "PTH",    # flake8-use-pathlib
  "PERF",   # Perflint
]
ignore = [
  "D415", # First line ending with punctuation - annoying for module docstrings
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Django-specific settings
[tool.ruff.lint.per-file-ignores]
# Ignore migrations directory - auto-generated files don't need linting
"*/migrations/*" = ["ALL"]

[tool.ruff.lint.pyflakes]
# Django ForeignKey with generic syntax (e.g., ForeignKey["User"])
extend-generics = ["django.db.models.ForeignKey"]

[tool.ruff.lint.isort]
# Django apps to treat as first-party imports
known-first-party = ["config", "users", "leagues", "services"]

[tool.ruff.lint.pydocstyle]
# Enforce Google-style docstrings for Sphinx compatibility
# This checks that docstrings follow Google style conventions
convention = "google"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks (e.g., .. code-block:: python) and doctests are all supported.
# Enabled since we occasionally include code examples in docstrings.
# This will become the default in future Ruff versions.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
